+++ src/main/java/org/prgrms/kdt/apis/Api.java
@@ -1,23 +1,17 @@
-package org.prgrms.kdt.api;
+package org.prgrms.kdt.apis;
 
 import static org.prgrms.kdt.util.LocalDateUtil.isValid;
-import static org.springframework.http.MediaType.*;
 
 import java.net.URI;
-import java.util.Arrays;
-import java.util.List;
 import javax.validation.Valid;
 import org.prgrms.kdt.customer.CustomerDto;
 import org.prgrms.kdt.customer.CustomerService;
 import org.prgrms.kdt.exception.BadRequestException;
-import org.prgrms.kdt.exception.ErrorResponse;
-import org.prgrms.kdt.form.VoucherForm;
 import org.prgrms.kdt.voucher.VoucherDto;
 import org.prgrms.kdt.voucher.VoucherService;
 import org.prgrms.kdt.voucher.VoucherType;
 import org.prgrms.kdt.wallet.WalletDto;
 import org.prgrms.kdt.wallet.WalletService;
-import org.springframework.http.MediaType;
 import org.springframework.http.ResponseEntity;
 import org.springframework.validation.BindingResult;
 import org.springframework.web.bind.annotation.DeleteMapping;
@@ -26,7 +20,6 @@ import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestMapping;
-import org.springframework.web.bind.annotation.RequestMethod;
 import org.springframework.web.bind.annotation.RequestParam;
 import org.springframework.web.bind.annotation.RestController;
 
@@ -36,8 +29,8 @@ import org.springframework.web.bind.annotation.RestController;
  */
 
 @RestController
-@RequestMapping(value = Apis.PRE_FIX)
-public class Apis {
+@RequestMapping(value = Api.PRE_FIX)
+public class Api {
 
     protected static final String PRE_FIX = "/kdt/api/v1";
     protected static final String WALLET = "/customers/wallet";
@@ -45,13 +38,12 @@ public class Apis {
     protected static final String VOUCHER = "/vouchers/{voucherId}";
     protected static final String VOUCHERS = "/vouchers";
     protected static final String SEARCH = "/search";
-    protected static final String VOUCHER_TYPE = "/type/{voucherType}";
 
     private final CustomerService customerService;
     private final VoucherService voucherService;
     private final WalletService walletService;
 
-    public Apis(CustomerService customerService, VoucherService voucherService, WalletService walletService) {
+    public Api(CustomerService customerService, VoucherService voucherService, WalletService walletService) {
         this.customerService = customerService;
         this.voucherService = voucherService;
         this.walletService = walletService;
@@ -109,15 +101,15 @@ public class Apis {
         return ResponseEntity.ok().build();
     }
 
-    @GetMapping(VOUCHERS + SEARCH + VOUCHER_TYPE)
-    public ResponseEntity getByVoucherType(@PathVariable String voucherType) {
+    @GetMapping(VOUCHERS + SEARCH + "/type")
+    public ResponseEntity getByVoucherType(@RequestParam String voucherType) {
         if (!VoucherType.isValue(voucherType)) {
             throw  new BadRequestException("not exist : " + voucherType);
         }
         return ResponseEntity.ok().body(voucherService.getVoucherByVoucherType(voucherType));
     }
 
-    @GetMapping(VOUCHERS + SEARCH)
+    @GetMapping(VOUCHERS + SEARCH + "/createAt")
     public ResponseEntity getByCreatedAt(@RequestParam String beforeDate,
                                          @RequestParam String afterDate) {
         if (!isValid(beforeDate)) {
